finished conditional logic

+++ lib/tasks/dev.rake
@@ -6,8 +6,23 @@ task({ :sample_data => :environment }) do
     ActiveRecord::Base.connection.reset_pk_sequence!(t)
   end
 
-  Board.destroy_all
-  Post.destroy_all
+  if Rails.env.development?
+    User.destroy_all
+    Board.destroy_all
+    Post.destroy_all
+  end
+
+
+  usernames = ["alice", "bob","carol", "dave", "eve"]
+
+  usernames.each do |username|
+
+    user= User.new
+    user.email = "#{username}@example.com"
+    user.password = "password"
+    user.save
+
+  end
   
   5.times do
     board = Board.new
